+++ app/src/main/res/navigation/nav_graph.xml
@@ -411,7 +411,7 @@
     </fragment>
     <fragment
         android:id="@+id/testFragment"
-        android:name="app.cashadvisor.authorization.presentation.ui.diagrams.TestFragment"
+        android:name="app.cashadvisor.authorization.presentation.ui.test.TestFragment"
         android:label="fragment_test"
         tools:layout="@layout/fragment_test" />
 </navigation>
